refactor(cli): extract agent, skill, hand, channel, and mcp command groups
Move the agent (spawn/list/chat/kill/new/set/sessions/message),
skill (install/list/test/publish/create/evolve/pending), hand,
channel, and `mcp <sub>` handlers out of main.rs into their own
commands/*.rs modules. PreparedAgentManifest moves with the agent
group. The new mcp handler module is named `mcp_cmds` to avoid
colliding with the existing stdio-server `mcp` module.

Pure code move, no behavior change.
