+++ crates/librefang-cli/src/main.rs
@@ -670,697 +670,6 @@ fn main() {
     }
 }
 
-struct PreparedAgentManifest {
-    manifest: AgentManifest,
-    manifest_toml: String,
-    source_label: String,
-}
-
-fn cmd_agent_spawn(
-    config: Option<PathBuf>,
-    manifest_path: PathBuf,
-    name_override: Option<String>,
-    dry_run: bool,
-) {
-    let prepared = prepared_agent_manifest_from_path(&manifest_path, name_override.as_deref());
-    if dry_run {
-        preview_agent_manifest(&prepared);
-        return;
-    }
-    spawn_prepared_agent(config, prepared);
-}
-
-fn cmd_spawn_alias(
-    config: Option<PathBuf>,
-    target: Option<String>,
-    template_path: Option<PathBuf>,
-    name_override: Option<String>,
-    dry_run: bool,
-) {
-    if template_path.is_some() && target.is_some() {
-        ui::error_with_fix(
-            "Choose either a positional target or `--template`, not both.",
-            "Use `librefang spawn coder` or `librefang spawn --template agents/custom/my-agent.toml`.",
-        );
-        std::process::exit(1);
-    }
-
-    if target.is_none() && template_path.is_none() {
-        if name_override.is_some() {
-            ui::error_with_fix(
-                "`--name` requires a template name or manifest path.",
-                "Use `librefang spawn coder --name backend-coder` or `librefang spawn --template path/to/agent.toml --name backend-coder`.",
-            );
-            std::process::exit(1);
-        }
-        if dry_run {
-            ui::error_with_fix(
-                "Dry run needs a template name or manifest path.",
-                "Use `librefang spawn coder --dry-run` or `librefang spawn --template path/to/agent.toml --dry-run`.",
-            );
-            std::process::exit(1);
-        }
-        cmd_agent_new(config, None);
-        return;
-    }
-
-    if let Some(path) = template_path {
-        let prepared = prepared_agent_manifest_from_path(&path, name_override.as_deref());
-        if dry_run {
-            preview_agent_manifest(&prepared);
-        } else {
-            spawn_prepared_agent(config, prepared);
-        }
-        return;
-    }
-
-    let target = target.expect("target checked above");
-    let manifest_path = PathBuf::from(&target);
-    if manifest_path.exists() {
-        let prepared = prepared_agent_manifest_from_path(&manifest_path, name_override.as_deref());
-        if dry_run {
-            preview_agent_manifest(&prepared);
-        } else {
-            spawn_prepared_agent(config, prepared);
-        }
-        return;
-    }
-
-    let templates = templates::load_all_templates();
-    let template = templates
-        .iter()
-        .find(|t| t.name == target)
-        .unwrap_or_else(|| {
-            ui::error_with_fix(
-                &format!("Template or manifest path not found: {target}"),
-                "Run `librefang agent new` to browse templates, or pass a valid manifest path.",
-            );
-            std::process::exit(1);
-        });
-    if dry_run {
-        let prepared = prepared_agent_manifest_from_template(template, name_override.as_deref());
-        preview_agent_manifest(&prepared);
-    } else {
-        spawn_template_agent(config, template, name_override.as_deref());
-    }
-}
-
-fn prepared_agent_manifest_from_path(
-    manifest_path: &std::path::Path,
-    name_override: Option<&str>,
-) -> PreparedAgentManifest {
-    if !manifest_path.exists() {
-        ui::error_with_fix(
-            &i18n::t_args(
-                "manifest-not-found",
-                &[("path", &manifest_path.display().to_string())],
-            ),
-            &i18n::t("manifest-not-found-fix"),
-        );
-        std::process::exit(1);
-    }
-
-    let contents = std::fs::read_to_string(manifest_path).unwrap_or_else(|e| {
-        eprintln!(
-            "{}",
-            i18n::t_args("error-reading-manifest", &[("error", &e.to_string())])
-        );
-        std::process::exit(1);
-    });
-
-    prepared_agent_manifest_from_contents(
-        &contents,
-        manifest_path.display().to_string(),
-        name_override,
-    )
-}
-
-fn prepared_agent_manifest_from_template(
-    template: &templates::AgentTemplate,
-    name_override: Option<&str>,
-) -> PreparedAgentManifest {
-    prepared_agent_manifest_from_contents(
-        &template.content,
-        format!("template:{}", template.name),
-        name_override,
-    )
-}
-
-fn prepared_agent_manifest_from_contents(
-    contents: &str,
-    source_label: String,
-    name_override: Option<&str>,
-) -> PreparedAgentManifest {
-    let mut manifest: AgentManifest = toml::from_str(contents).unwrap_or_else(|e| {
-        ui::error_with_fix(
-            &format!("Failed to parse agent manifest from {source_label}: {e}"),
-            "Check the manifest TOML syntax and required fields.",
-        );
-        std::process::exit(1);
-    });
-
-    if let Some(name) = name_override {
-        manifest.name = name.to_string();
-    }
-
-    let manifest_toml = if name_override.is_some() {
-        toml::to_string_pretty(&manifest).unwrap_or_else(|e| {
-            ui::error(&format!("Failed to serialize updated manifest: {e}"));
-            std::process::exit(1);
-        })
-    } else {
-        contents.to_string()
-    };
-
-    PreparedAgentManifest {
-        manifest,
-        manifest_toml,
-        source_label,
-    }
-}
-
-fn preview_agent_manifest(prepared: &PreparedAgentManifest) {
-    ui::section("Agent Dry Run");
-    ui::kv("Source", &prepared.source_label);
-    ui::kv("Name", &prepared.manifest.name);
-    ui::kv("Version", &prepared.manifest.version);
-    ui::kv("Module", &prepared.manifest.module);
-    ui::kv(
-        "Model",
-        &format!(
-            "{}/{}",
-            prepared.manifest.model.provider, prepared.manifest.model.model
-        ),
-    );
-    ui::kv(
-        "Tools",
-        &prepared.manifest.capabilities.tools.len().to_string(),
-    );
-    ui::kv("Skills", &prepared.manifest.skills.len().to_string());
-    if !prepared.manifest.tags.is_empty() {
-        ui::kv("Tags", &prepared.manifest.tags.join(", "));
-    }
-    if !prepared.manifest.description.is_empty() {
-        ui::kv("Description", &prepared.manifest.description);
-    }
-    ui::success("Manifest parsed successfully. No agent was spawned.");
-}
-
-fn spawn_prepared_agent(config: Option<PathBuf>, prepared: PreparedAgentManifest) {
-    if let Some(base) = find_daemon() {
-        let client = daemon_client();
-        let body = daemon_json(
-            client
-                .post(format!("{base}/api/agents"))
-                .json(&serde_json::json!({"manifest_toml": prepared.manifest_toml}))
-                .send(),
-        );
-        if body.get("agent_id").is_some() {
-            println!("{}", i18n::t("agent-spawn-success"));
-            println!("  ID:   {}", body["agent_id"].as_str().unwrap_or("?"));
-            println!(
-                "  Name: {}",
-                body["name"]
-                    .as_str()
-                    .unwrap_or(prepared.manifest.name.as_str())
-            );
-        } else {
-            eprintln!(
-                "{}",
-                i18n::t_args(
-                    "agent-spawn-agent-failed",
-                    &[("error", body["error"].as_str().unwrap_or("Unknown error"))]
-                )
-            );
-            std::process::exit(1);
-        }
-    } else {
-        let agent_name = prepared.manifest.name.clone();
-        let kernel = boot_kernel(config);
-        match kernel.spawn_agent_with_source(prepared.manifest, None) {
-            Ok(id) => {
-                println!("{}", i18n::t("agent-spawn-inprocess-mode"));
-                println!("  ID:   {id}");
-                println!("  Name: {agent_name}");
-                println!("\n  {}", i18n::t("agent-note-lost"));
-                println!("  {}", i18n::t("agent-note-persistent"));
-            }
-            Err(e) => {
-                eprintln!(
-                    "{}",
-                    i18n::t_args("agent-spawn-agent-failed", &[("error", &e.to_string())])
-                );
-                std::process::exit(1);
-            }
-        }
-    }
-}
-
-fn cmd_agent_list(config: Option<PathBuf>, json: bool) {
-    if let Some(base) = find_daemon() {
-        let client = daemon_client();
-        let body = daemon_json(client.get(format!("{base}/api/agents")).send());
-
-        if json {
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&body).unwrap_or_default()
-            );
-            return;
-        }
-
-        let agents = body
-            .get("items")
-            .and_then(|v| v.as_array())
-            .or_else(|| body.as_array());
-
-        match agents {
-            Some(agents) if agents.is_empty() => println!("{}", i18n::t("agent-no-agents")),
-            Some(agents) => {
-                // Render via the shared Table builder so column widths
-                // self-size to the actual content (instead of hard-coded
-                // {:<38} which truncates / over-pads), and so piped output
-                // automatically falls back to ASCII (#3306).
-                let mut t = crate::table::Table::new(&["ID", "NAME", "STATE", "PROVIDER", "MODEL"]);
-                for a in agents {
-                    t.add_row(&[
-                        a["id"].as_str().unwrap_or("?"),
-                        a["name"].as_str().unwrap_or("?"),
-                        a["state"].as_str().unwrap_or("?"),
-                        a["model_provider"].as_str().unwrap_or("?"),
-                        a["model_name"].as_str().unwrap_or("?"),
-                    ]);
-                }
-                t.print();
-            }
-            None => println!("{}", i18n::t("agent-no-agents")),
-        }
-    } else {
-        let kernel = boot_kernel(config);
-        let agents = kernel.agent_registry_ref().list();
-
-        if json {
-            let list: Vec<serde_json::Value> = agents
-                .iter()
-                .map(|e| {
-                    serde_json::json!({
-                        "id": e.id.to_string(),
-                        "name": e.name,
-                        "state": format!("{:?}", e.state),
-                        "created_at": e.created_at.to_rfc3339(),
-                    })
-                })
-                .collect();
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&list).unwrap_or_default()
-            );
-            return;
-        }
-
-        if agents.is_empty() {
-            println!("{}", i18n::t("agent-no-agents"));
-            return;
-        }
-
-        let mut t = crate::table::Table::new(&["ID", "NAME", "STATE", "CREATED"]);
-        for entry in agents {
-            let id = entry.id.to_string();
-            let state = format!("{:?}", entry.state);
-            let created = entry.created_at.format("%Y-%m-%d %H:%M").to_string();
-            t.add_row(&[
-                id.as_str(),
-                entry.name.as_str(),
-                state.as_str(),
-                created.as_str(),
-            ]);
-        }
-        t.print();
-    }
-}
-
-fn cmd_agent_chat(config: Option<PathBuf>, agent_id_str: &str) {
-    ensure_initialized(&config);
-    tui::chat_runner::run_chat_tui(config, Some(agent_id_str.to_string()));
-}
-
-fn cmd_agent_kill(config: Option<PathBuf>, agent_id_str: &str) {
-    if let Some(base) = find_daemon() {
-        let agent_id = resolve_agent_id(&base, agent_id_str);
-        let client = daemon_client();
-        // Refs #4614: explicit `librefang agent kill <id>` IS the user's
-        // confirmation. The API requires `?confirm=true` on DELETE so the
-        // canonical UUID is purged on the kill (matching the issue's
-        // "explicit delete" semantics). Internal lifecycle resets call
-        // `kernel.kill_agent` directly and skip this path.
-        let body = daemon_json(
-            client
-                .delete(format!("{base}/api/agents/{agent_id}?confirm=true"))
-                .send(),
-        );
-        if body.get("status").is_some() {
-            println!("{}", i18n::t_args("agent-killed", &[("id", &agent_id)]));
-        } else {
-            eprintln!(
-                "{}",
-                i18n::t_args(
-                    "agent-kill-failed",
-                    &[("error", body["error"].as_str().unwrap_or("Unknown error"))]
-                )
-            );
-            std::process::exit(1);
-        }
-    } else {
-        let agent_id: AgentId = agent_id_str.parse().unwrap_or_else(|_| {
-            eprintln!(
-                "{}",
-                i18n::t_args("agent-invalid-id", &[("id", agent_id_str)])
-            );
-            std::process::exit(1);
-        });
-        let kernel = boot_kernel(config);
-        // Direct-kernel path (no daemon): mirror the API's confirmed-delete
-        // semantics so behavior matches whether the daemon is running or not.
-        match kernel.kill_agent_with_purge(agent_id, true) {
-            Ok(()) => println!(
-                "{}",
-                i18n::t_args("agent-killed", &[("id", &agent_id.to_string())])
-            ),
-            Err(e) => {
-                eprintln!(
-                    "{}",
-                    i18n::t_args("agent-kill-failed", &[("error", &e.to_string())])
-                );
-                std::process::exit(1);
-            }
-        }
-    }
-}
-
-/// Refs #4614 — `librefang agent delete <name>` with confirmation prompt.
-///
-/// Looks up the canonical UUID for `name` via `GET /api/agents/identities`
-/// (or directly from the kernel registry when no daemon is running),
-/// prints the destructive-action warning, and either prompts `[y/N]` or
-/// proceeds immediately when `--yes` is set. Then issues the confirmed
-/// DELETE. This is the long-form companion to `librefang agent kill <id>`
-/// — useful when the operator only knows the agent's name.
-fn cmd_agent_delete(config: Option<PathBuf>, name: &str, yes: bool) {
-    eprintln!("WARNING: Deleting agent \"{name}\" will permanently remove its canonical UUID");
-    eprintln!("    and all associated memories and sessions.");
-    eprintln!("    This action cannot be undone.");
-    if !yes && !prompt_yes_no("Confirm?", false) {
-        eprintln!("Aborted.");
-        std::process::exit(1);
-    }
-
-    if let Some(base) = find_daemon() {
-        let client = daemon_client();
-        // Resolve name → UUID via the identity registry endpoint.
-        let canonical_uuid = match lookup_canonical_uuid(&base, name) {
-            Some(id) => id,
-            None => {
-                eprintln!(
-                    "No canonical UUID recorded for agent name '{name}' — nothing to delete."
-                );
-                std::process::exit(1);
-            }
-        };
-        let body = daemon_json(
-            client
-                .delete(format!("{base}/api/agents/{canonical_uuid}?confirm=true"))
-                .send(),
-        );
-        if body.get("status").is_some() {
-            println!("Agent \"{name}\" deleted (canonical UUID purged).");
-        } else {
-            eprintln!(
-                "Failed to delete agent: {}",
-                body["error"].as_str().unwrap_or("Unknown error")
-            );
-            std::process::exit(1);
-        }
-    } else {
-        let kernel = boot_kernel(config);
-        let canonical_uuid = match kernel.identities_ref().get(name) {
-            Some(id) => id,
-            None => {
-                eprintln!(
-                    "No canonical UUID recorded for agent name '{name}' — nothing to delete."
-                );
-                std::process::exit(1);
-            }
-        };
-        match kernel.kill_agent_with_purge(canonical_uuid, true) {
-            Ok(()) => println!("Agent \"{name}\" deleted (canonical UUID purged)."),
-            Err(e) => {
-                eprintln!("Failed to delete agent: {e}");
-                std::process::exit(1);
-            }
-        }
-    }
-}
-
-/// Refs #4614 — `librefang agent reset-uuid <name>` with confirmation.
-///
-/// Drops the canonical UUID binding without killing a running agent. The
-/// next spawn under `name` re-derives a fresh UUID and registers it as
-/// the new canonical binding; prior sessions / memories tied to the old
-/// UUID are orphaned. `--yes` skips the prompt.
-fn cmd_agent_reset_uuid(config: Option<PathBuf>, name: &str, yes: bool) {
-    eprintln!("WARNING: Resetting the canonical UUID for \"{name}\" will orphan all sessions");
-    eprintln!("    and memories tied to its current UUID. The next spawn under this");
-    eprintln!("    name will start with a fresh UUID. This action cannot be undone.");
-    if !yes && !prompt_yes_no("Confirm?", false) {
-        eprintln!("Aborted.");
-        std::process::exit(1);
-    }
-
-    if let Some(base) = find_daemon() {
-        let client = daemon_client();
-        let body = daemon_json(
-            client
-                .post(format!(
-                    "{base}/api/agents/identities/{}/reset",
-                    percent_encode_path_segment(name)
-                ))
-                .query(&[("confirm", "true")])
-                .send(),
-        );
-        if body.get("status").is_some() {
-            println!(
-                "Canonical UUID for \"{name}\" reset (was {}).",
-                body["previous_canonical_uuid"]
-                    .as_str()
-                    .unwrap_or("<unknown>")
-            );
-        } else {
-            eprintln!(
-                "Failed to reset canonical UUID: {}",
-                body["error"].as_str().unwrap_or("Unknown error")
-            );
-            std::process::exit(1);
-        }
-    } else {
-        let kernel = boot_kernel(config);
-        match kernel.identities_ref().purge(name) {
-            Some(prev) => println!("Canonical UUID for \"{name}\" reset (was {prev})."),
-            None => {
-                eprintln!("No canonical UUID recorded for agent name '{name}'.");
-                std::process::exit(1);
-            }
-        }
-    }
-}
-
-/// Refs #4614 — `librefang agent merge-history` placeholder.
-///
-/// The cross-table reassignment is not yet implemented — see the
-/// long_about on `AgentCommands::MergeHistory` for the rationale (deep
-/// memory-substrate surgery across 10+ tables under one transaction).
-fn cmd_agent_merge_history(name: &str, from: &str) {
-    eprintln!("merge-history is not yet implemented (refs #4614 follow-up).");
-    eprintln!("Reassignment of sessions / memories from {from} to the canonical UUID");
-    eprintln!("for agent \"{name}\" requires cross-table SQL surgery in the memory");
-    eprintln!("substrate that is being tracked separately.");
-    std::process::exit(2);
-}
-
-/// Look up the canonical UUID for `name` via the identity-registry
-/// endpoint. Returns `None` if no entry exists (or on any HTTP error —
-/// the caller surfaces a friendly message).
-fn lookup_canonical_uuid(base: &str, name: &str) -> Option<String> {
-    let client = daemon_client();
-    let resp = client
-        .get(format!("{base}/api/agents/identities"))
-        .send()
-        .ok()?;
-    let entries: serde_json::Value = resp.json().ok()?;
-    let arr = entries.as_array()?;
-    for entry in arr {
-        if entry["name"].as_str() == Some(name) {
-            return entry["canonical_uuid"].as_str().map(String::from);
-        }
-    }
-    None
-}
-
-fn cmd_agent_set(agent_id_str: &str, field: &str, value: &str) {
-    match field {
-        "model" => {
-            if let Some(base) = find_daemon() {
-                let agent_id = resolve_agent_id(&base, agent_id_str);
-                let client = daemon_client();
-                let body = daemon_json(
-                    client
-                        .put(format!("{base}/api/agents/{agent_id}/model"))
-                        .json(&serde_json::json!({"model": value}))
-                        .send(),
-                );
-                if body.get("status").is_some() {
-                    println!("Agent {agent_id} model set to {value}.");
-                } else {
-                    eprintln!(
-                        "Failed to set model: {}",
-                        body["error"].as_str().unwrap_or("Unknown error")
-                    );
-                    std::process::exit(1);
-                }
-            } else {
-                eprintln!("No running daemon found. Start one with: librefang start");
-                std::process::exit(1);
-            }
-        }
-        _ => {
-            eprintln!("Unknown field: {field}. Supported fields: model");
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_agent_new(config: Option<PathBuf>, template_name: Option<String>) {
-    let all_templates = templates::load_all_templates();
-    if all_templates.is_empty() {
-        ui::error_with_fix(
-            "No agent templates found",
-            "Run `librefang init` to set up the agents directory",
-        );
-        std::process::exit(1);
-    }
-
-    // Resolve template: by name or interactive picker
-    let chosen = match template_name {
-        Some(ref name) => match all_templates.iter().find(|t| t.name == *name) {
-            Some(t) => t,
-            None => {
-                ui::error_with_fix(
-                    &format!("Template '{name}' not found"),
-                    "Run `librefang agent new` to see available templates",
-                );
-                std::process::exit(1);
-            }
-        },
-        None => {
-            ui::section(&i18n::t("section-agent-templates"));
-            ui::blank();
-            for (i, t) in all_templates.iter().enumerate() {
-                let desc = if t.description.is_empty() {
-                    String::new()
-                } else {
-                    format!("  {}", t.description)
-                };
-                println!(
-                    "    {:>2}. {:<22}{}",
-                    i + 1,
-                    t.name,
-                    colored::Colorize::dimmed(desc.as_str())
-                );
-            }
-            ui::blank();
-            let choice = prompt_input("  Choose template [1]: ");
-            let idx = if choice.is_empty() {
-                0
-            } else {
-                choice
-                    .parse::<usize>()
-                    .unwrap_or(1)
-                    .saturating_sub(1)
-                    .min(all_templates.len() - 1)
-            };
-            &all_templates[idx]
-        }
-    };
-
-    // Spawn the agent
-    spawn_template_agent(config, chosen, None);
-}
-
-/// Spawn an agent from a template, via daemon or in-process.
-fn spawn_template_agent(
-    config: Option<PathBuf>,
-    template: &templates::AgentTemplate,
-    name_override: Option<&str>,
-) {
-    let prepared = prepared_agent_manifest_from_template(template, name_override);
-    let agent_name = prepared.manifest.name.clone();
-
-    if let Some(base) = find_daemon() {
-        let client = daemon_client();
-        let body = daemon_json(
-            client
-                .post(format!("{base}/api/agents"))
-                .json(&serde_json::json!({"manifest_toml": prepared.manifest_toml}))
-                .send(),
-        );
-        if let Some(id) = body["agent_id"].as_str() {
-            ui::blank();
-            ui::success(&i18n::t_args("agent-spawned", &[("name", &agent_name)]));
-            ui::kv(&i18n::t("label-id"), id);
-            if let Some(model) = body["model_name"].as_str() {
-                let provider = body["model_provider"].as_str().unwrap_or("?");
-                ui::kv(&i18n::t("label-model"), &format!("{provider}/{model}"));
-            }
-            ui::blank();
-            ui::hint(&i18n::t_args(
-                "hint-chat-with-agent",
-                &[("name", &agent_name)],
-            ));
-        } else {
-            ui::error(&i18n::t_args(
-                "agent-spawn-failed",
-                &[("error", body["error"].as_str().unwrap_or("Unknown error"))],
-            ));
-            std::process::exit(1);
-        }
-    } else {
-        let kernel = boot_kernel(config);
-        match kernel.spawn_agent(prepared.manifest) {
-            Ok(id) => {
-                ui::blank();
-                ui::success(&i18n::t_args(
-                    "agent-spawned-inprocess",
-                    &[("name", &agent_name)],
-                ));
-                ui::kv(&i18n::t("label-id"), &id.to_string());
-                ui::blank();
-                ui::hint(&i18n::t_args(
-                    "hint-chat-with-agent",
-                    &[("name", &agent_name)],
-                ));
-                ui::hint(&i18n::t("hint-agent-lost-on-exit"));
-                ui::hint(&i18n::t("hint-persistent-agents"));
-            }
-            Err(e) => {
-                ui::error(&i18n::t_args(
-                    "agent-spawn-agent-failed",
-                    &[("error", &e.to_string())],
-                ));
-                std::process::exit(1);
-            }
-        }
-    }
-}
-
 fn cmd_doctor(json: bool, repair: bool) {
     // BrokenPipe protection for the WHOLE command, not just the --json
     // branch. `librefang doctor | head -5` and similar pipelines drop the
@@ -2964,1831 +2273,6 @@ fn cmd_migrate(args: MigrateArgs) {
 }
 
 // ---------------------------------------------------------------------------
-// Skill commands
-// ---------------------------------------------------------------------------
-
-/// Resolve the skills directory: global or per-hand workspace.
-fn resolve_skills_dir(hand: Option<&str>) -> PathBuf {
-    let home = librefang_home();
-    match hand {
-        None => home.join("skills"),
-        Some(hand_id) => {
-            let hand_dir = home.join("workspaces").join("hands").join(hand_id);
-            if !hand_dir.exists() {
-                eprintln!("Hand '{hand_id}' not found at {}", hand_dir.display());
-                std::process::exit(1);
-            }
-            hand_dir.join("skills")
-        }
-    }
-}
-
-fn cmd_skill_install(source: &str, hand: Option<&str>) {
-    let skills_dir = resolve_skills_dir(hand);
-    std::fs::create_dir_all(&skills_dir).unwrap_or_else(|e| {
-        eprintln!("Error creating skills directory: {e}");
-        std::process::exit(1);
-    });
-
-    let source_path = PathBuf::from(source);
-    if source_path.exists() && source_path.is_dir() {
-        // Local directory install
-        let manifest_path = source_path.join("skill.toml");
-        if !manifest_path.exists() {
-            // Check if it's an OpenClaw skill
-            if librefang_skills::openclaw_compat::detect_openclaw_skill(&source_path) {
-                println!("Detected OpenClaw skill format. Converting...");
-                match librefang_skills::openclaw_compat::convert_openclaw_skill(&source_path) {
-                    Ok(manifest) => {
-                        let dest = skills_dir.join(&manifest.skill.name);
-                        // Copy skill directory
-                        copy_dir_recursive(&source_path, &dest);
-                        if let Err(e) = librefang_skills::openclaw_compat::write_librefang_manifest(
-                            &dest, &manifest,
-                        ) {
-                            eprintln!("Failed to write manifest: {e}");
-                            std::process::exit(1);
-                        }
-                        if let Some(h) = hand {
-                            println!(
-                                "Installed OpenClaw skill '{}' to hand '{h}'",
-                                manifest.skill.name
-                            );
-                        } else {
-                            println!("Installed OpenClaw skill: {}", manifest.skill.name);
-                        }
-                    }
-                    Err(e) => {
-                        eprintln!("Failed to convert OpenClaw skill: {e}");
-                        std::process::exit(1);
-                    }
-                }
-                return;
-            }
-            eprintln!("No skill.toml found in {source}");
-            std::process::exit(1);
-        }
-
-        // Read manifest to get skill name
-        let toml_str = std::fs::read_to_string(&manifest_path).unwrap_or_else(|e| {
-            eprintln!("Error reading skill.toml: {e}");
-            std::process::exit(1);
-        });
-        let manifest: librefang_skills::SkillManifest =
-            toml::from_str(&toml_str).unwrap_or_else(|e| {
-                eprintln!("Error parsing skill.toml: {e}");
-                std::process::exit(1);
-            });
-
-        let dest = skills_dir.join(&manifest.skill.name);
-        copy_dir_recursive(&source_path, &dest);
-        if let Some(h) = hand {
-            println!(
-                "Installed skill '{}' v{} to hand '{h}'",
-                manifest.skill.name, manifest.skill.version
-            );
-        } else {
-            println!(
-                "Installed skill: {} v{}",
-                manifest.skill.name, manifest.skill.version
-            );
-        }
-    } else {
-        // Remote install from FangHub
-        let mut sp = progress::auto(&format!("Installing {source}"), None);
-        sp.tick(1);
-        let rt = tokio::runtime::Runtime::new().unwrap();
-        let client = librefang_skills::marketplace::MarketplaceClient::new(
-            librefang_skills::marketplace::MarketplaceConfig::default(),
-        );
-        match rt.block_on(client.install(source, &skills_dir)) {
-            Ok(version) => {
-                if let Some(h) = hand {
-                    sp.finish(&format!("Installed {source} {version} to hand '{h}'"));
-                } else {
-                    sp.finish(&format!("Installed {source} {version}"));
-                }
-            }
-            Err(e) => {
-                sp.finish_with_failure(&format!("Failed to install skill: {e}"));
-                std::process::exit(1);
-            }
-        }
-    }
-}
-
-fn cmd_skill_list(hand: Option<&str>) {
-    let skills_dir = resolve_skills_dir(hand);
-
-    let mut registry = librefang_skills::registry::SkillRegistry::new(skills_dir);
-    match registry.load_all() {
-        Ok(0) => {
-            if let Some(h) = hand {
-                println!("No skills installed for hand '{h}'.");
-            } else {
-                println!("No skills installed.");
-            }
-        }
-        Ok(count) => {
-            if let Some(h) = hand {
-                println!("{count} skill(s) installed for hand '{h}':\n");
-            } else {
-                println!("{count} skill(s) installed:\n");
-            }
-            let mut t = crate::table::Table::new(&["NAME", "VERSION", "TOOLS", "DESCRIPTION"]);
-            for skill in registry.list() {
-                t.add_row(&[
-                    &skill.manifest.skill.name,
-                    &skill.manifest.skill.version,
-                    &skill.manifest.tools.provided.len().to_string(),
-                    &skill.manifest.skill.description,
-                ]);
-            }
-            t.print();
-        }
-        Err(e) => {
-            eprintln!("Error loading skills: {e}");
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_skill_remove(name: &str, hand: Option<&str>) {
-    // Route through the safe uninstall path (lock + path-traversal
-    // guard) instead of `registry.remove()` which calls `remove_dir_all`
-    // with no serialisation against concurrent evolve operations.
-    let skills_dir = resolve_skills_dir(hand);
-    match librefang_skills::evolution::uninstall_skill(&skills_dir, name) {
-        Ok(_) => {
-            if let Some(h) = hand {
-                println!("Removed skill '{name}' from hand '{h}'");
-            } else {
-                println!("Removed skill: {name}");
-            }
-        }
-        Err(e) => {
-            eprintln!("Failed to remove skill: {e}");
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_skill_search(query: &str) {
-    let rt = tokio::runtime::Runtime::new().unwrap();
-    let client = librefang_skills::marketplace::MarketplaceClient::new(
-        librefang_skills::marketplace::MarketplaceConfig::default(),
-    );
-    match rt.block_on(client.search(query)) {
-        Ok(results) if results.is_empty() => println!("No skills found for \"{query}\"."),
-        Ok(results) => {
-            println!("Skills matching \"{query}\":\n");
-            for r in results {
-                println!("  {} ({})", r.name, r.stars);
-                if !r.description.is_empty() {
-                    println!("    {}", r.description);
-                }
-                println!("    {}", r.url);
-                println!();
-            }
-        }
-        Err(e) => {
-            eprintln!("Search failed: {e}");
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_skill_test(path: Option<PathBuf>, tool: Option<String>, input: Option<String>) {
-    let skill_path = resolve_skill_path(path);
-    let prepared =
-        librefang_skills::publish::prepare_local_skill(&skill_path).unwrap_or_else(|e| {
-            eprintln!("Skill validation failed: {e}");
-            std::process::exit(1);
-        });
-
-    println!(
-        "Validated skill: {} v{}",
-        prepared.manifest.skill.name, prepared.manifest.skill.version
-    );
-    println!(
-        "  Runtime: {:?}\n  Source: {}",
-        prepared.manifest.runtime.runtime_type,
-        prepared.source_dir.display()
-    );
-    if !prepared.manifest.skill.description.is_empty() {
-        println!("  Description: {}", prepared.manifest.skill.description);
-    }
-    if !prepared.manifest.tools.provided.is_empty() {
-        println!(
-            "  Tools: {}",
-            prepared
-                .manifest
-                .tools
-                .provided
-                .iter()
-                .map(|tool| tool.name.as_str())
-                .collect::<Vec<_>>()
-                .join(", ")
-        );
-    }
-    print_skill_warnings(&prepared.warnings);
-
-    if prepared.has_critical_warnings() {
-        eprintln!("Refusing to execute a skill with critical validation warnings.");
-        std::process::exit(1);
-    }
-
-    let Some(tool_name) = tool.or_else(|| {
-        prepared
-            .manifest
-            .tools
-            .provided
-            .first()
-            .map(|tool| tool.name.clone())
-    }) else {
-        println!("Validation only: no tool declared to execute.");
-        return;
-    };
-
-    let input_json = match input {
-        Some(input) => serde_json::from_str::<serde_json::Value>(&input).unwrap_or_else(|err| {
-            eprintln!("Invalid --input JSON: {err}");
-            std::process::exit(1);
-        }),
-        None => serde_json::json!({}),
-    };
-
-    let rt = tokio::runtime::Runtime::new().unwrap();
-    let result = if prepared.manifest.runtime.runtime_type == librefang_skills::SkillRuntime::Wasm {
-        // WASM skills execute in the real sandbox. We pass no kernel handle:
-        // pure-compute tools run end to end, while capability-bearing host
-        // calls return an error in the result rather than crashing — the right
-        // behaviour for a local smoke test outside a running daemon.
-        rt.block_on(librefang_runtime::tool_runner::execute_wasm_skill(
-            &prepared.manifest,
-            &prepared.source_dir,
-            &tool_name,
-            &input_json,
-            None,
-            "cli-test",
-        ))
-    } else {
-        let env_policy = load_skill_env_policy_from_config();
-        rt.block_on(librefang_skills::loader::execute_skill_tool(
-            &prepared.manifest,
-            &prepared.source_dir,
-            &tool_name,
-            &input_json,
-            env_policy.as_ref(),
-        ))
-    };
-    match result {
-        Ok(result) => {
-            println!("\nTool result ({tool_name}):");
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&result.output).unwrap_or_default()
-            );
-            if result.is_error {
-                std::process::exit(1);
-            }
-        }
-        Err(librefang_skills::SkillError::RuntimeNotAvailable(message)) => {
-            println!("\nValidation complete.");
-            println!("Execution skipped: {message}");
-        }
-        Err(err) => {
-            eprintln!("Skill execution failed: {err}");
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_skill_publish(
-    path: Option<PathBuf>,
-    repo: Option<String>,
-    tag: Option<String>,
-    output: Option<PathBuf>,
-    dry_run: bool,
-) {
-    let skill_path = resolve_skill_path(path);
-    let prepared =
-        librefang_skills::publish::prepare_local_skill(&skill_path).unwrap_or_else(|e| {
-            eprintln!("Skill validation failed: {e}");
-            std::process::exit(1);
-        });
-
-    println!(
-        "Preparing skill: {} v{}",
-        prepared.manifest.skill.name, prepared.manifest.skill.version
-    );
-    print_skill_warnings(&prepared.warnings);
-    if prepared.has_critical_warnings() {
-        eprintln!("Refusing to publish a skill with critical validation warnings.");
-        std::process::exit(1);
-    }
-
-    let output_dir = output.unwrap_or_else(|| prepared.source_dir.join("dist"));
-    let packaged = librefang_skills::publish::package_prepared_skill(&prepared, &output_dir)
-        .unwrap_or_else(|e| {
-            eprintln!("Failed to package skill: {e}");
-            std::process::exit(1);
-        });
-
-    println!(
-        "Bundle created: {}\n  SHA256: {}\n  Size: {} bytes",
-        packaged.archive_path.display(),
-        packaged.sha256,
-        packaged.size_bytes
-    );
-
-    let repo = repo.unwrap_or_else(|| format!("librefang-skills/{}", packaged.manifest.skill.name));
-    let tag = tag.unwrap_or_else(|| format!("v{}", packaged.manifest.skill.version));
-
-    if dry_run {
-        println!("Dry run only.");
-        println!("  Repo: {repo}\n  Tag: {tag}");
-        return;
-    }
-
-    let token = std::env::var("GITHUB_TOKEN")
-        .or_else(|_| std::env::var("GH_TOKEN"))
-        .unwrap_or_else(|_| {
-            eprintln!("Set GITHUB_TOKEN or GH_TOKEN to publish, or re-run with --dry-run.");
-            std::process::exit(1);
-        });
-
-    let release_notes = format!(
-        "{}\n\nSHA256: `{}`\n\nInstall with:\n`librefang skill install {}`",
-        packaged.manifest.skill.description, packaged.sha256, packaged.manifest.skill.name
-    );
-    let release_name = format!(
-        "{} {}",
-        packaged.manifest.skill.name, packaged.manifest.skill.version
-    );
-
-    let mut sp = progress::auto(
-        &format!("Publishing {}@{tag}", packaged.manifest.skill.name),
-        None,
-    );
-    sp.tick(1);
-    let rt = tokio::runtime::Runtime::new().unwrap();
-    let client = librefang_skills::marketplace::MarketplaceClient::new(
-        librefang_skills::marketplace::MarketplaceConfig::default(),
-    );
-    let published = rt
-        .block_on(
-            client.publish_bundle(librefang_skills::marketplace::MarketplacePublishRequest {
-                repo: &repo,
-                tag: &tag,
-                bundle_path: &packaged.archive_path,
-                release_name: &release_name,
-                release_notes: &release_notes,
-                token: &token,
-            }),
-        )
-        .unwrap_or_else(|e| {
-            sp.finish_with_failure(&format!("Publish failed: {e}"));
-            std::process::exit(1);
-        });
-
-    sp.finish(&format!(
-        "Published {} to {}@{}",
-        published.asset_name, published.repo, published.tag
-    ));
-    if !published.html_url.is_empty() {
-        println!("Release: {}", published.html_url);
-    }
-}
-
-fn resolve_skill_path(path: Option<PathBuf>) -> PathBuf {
-    path.unwrap_or_else(|| {
-        std::env::current_dir().unwrap_or_else(|e| {
-            eprintln!("Could not determine current directory: {e}");
-            std::process::exit(1);
-        })
-    })
-}
-
-fn print_skill_warnings(warnings: &[librefang_skills::verify::SkillWarning]) {
-    if warnings.is_empty() {
-        println!("  Warnings: none");
-        return;
-    }
-
-    println!("  Warnings:");
-    for warning in warnings {
-        println!(
-            "    [{}] {}",
-            severity_label(warning.severity),
-            warning.message
-        );
-    }
-}
-
-fn severity_label(severity: librefang_skills::verify::WarningSeverity) -> &'static str {
-    match severity {
-        librefang_skills::verify::WarningSeverity::Info => "info",
-        librefang_skills::verify::WarningSeverity::Warning => "warn",
-        librefang_skills::verify::WarningSeverity::Critical => "critical",
-    }
-}
-
-fn cmd_skill_create() {
-    let name = prompt_input("Skill name: ");
-    let description = prompt_input("Description: ");
-    let runtime = prompt_input("Runtime (python/node/wasm) [python]: ");
-    let runtime = if runtime.is_empty() {
-        "python".to_string()
-    } else {
-        runtime
-    };
-
-    let home = librefang_home();
-    let skill_dir = home.join("skills").join(&name);
-    std::fs::create_dir_all(skill_dir.join("src")).unwrap_or_else(|e| {
-        eprintln!("Error creating skill directory: {e}");
-        std::process::exit(1);
-    });
-
-    let tool_name = name.replace('-', "_");
-
-    // A Cargo package name must be `[A-Za-z0-9_-]+` and not start with a digit;
-    // a skill name can be anything the user typed. Derive a legal package name
-    // for the WASM scaffold's Cargo.toml. The artifact name is fixed to
-    // `skill` via `[lib] name`, so this only needs to be valid, not meaningful.
-    let pkg_name = {
-        let cleaned: String = name
-            .chars()
-            .map(|c| {
-                if c.is_ascii_alphanumeric() || c == '-' || c == '_' {
-                    c.to_ascii_lowercase()
-                } else {
-                    '-'
-                }
-            })
-            .collect();
-        let cleaned = cleaned.trim_matches('-');
-        if cleaned.is_empty() {
-            "skill".to_string()
-        } else if cleaned.starts_with(|c: char| c.is_ascii_digit()) {
-            format!("skill-{cleaned}")
-        } else {
-            cleaned.to_string()
-        }
-    };
-
-    // Per-runtime scaffold: the manifest `entry` path, the files to write
-    // (relative to the skill dir), and any extra build steps the author must
-    // run before the entry exists.
-    struct Scaffold {
-        entry: String,
-        files: Vec<(String, String)>,
-        build_steps: Vec<String>,
-    }
-
-    let scaffold = match runtime.as_str() {
-        "python" => Scaffold {
-            entry: "src/main.py".to_string(),
-            files: vec![(
-                "src/main.py".to_string(),
-                format!(
-                    r#"#!/usr/bin/env python3
-"""LibreFang skill: {name}"""
-import json
-import sys
-
-def main():
-    payload = json.loads(sys.stdin.read())
-    tool_name = payload["tool"]
-    input_data = payload["input"]
-
-    # TODO: Implement your skill logic here
-    result = {{"result": f"Processed: {{input_data.get('input', '')}}"}}
-
-    print(json.dumps(result))
-
-if __name__ == "__main__":
-    main()
-"#
-                ),
-            )],
-            build_steps: vec![],
-        },
-        "node" => Scaffold {
-            entry: "src/index.js".to_string(),
-            files: vec![(
-                "src/index.js".to_string(),
-                format!(
-                    r#"// LibreFang skill: {name}
-const chunks = [];
-process.stdin.on("data", (c) => chunks.push(c));
-process.stdin.on("end", () => {{
-  const payload = JSON.parse(Buffer.concat(chunks).toString());
-  const input = payload.input || {{}};
-  // TODO: Implement your skill logic here
-  const result = {{ result: `Processed: ${{input.input ?? ""}}` }};
-  process.stdout.write(JSON.stringify(result));
-}});
-"#
-                ),
-            )],
-            build_steps: vec![],
-        },
-        "wasm" => Scaffold {
-            // Entry is the artifact at the skill root, NOT under target/: the
-            // packager (`should_include_entry`) excludes `target/`, so a skill
-            // referencing the build dir would publish without its binary. The
-            // build step copies the compiled module to the root.
-            entry: "skill.wasm".to_string(),
-            files: vec![
-                (
-                    "Cargo.toml".to_string(),
-                    format!(
-                        r#"[package]
-name = "{pkg_name}"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-# Fixed name so the artifact is always `skill.wasm` regardless of package name.
-name = "skill"
-crate-type = ["cdylib"]
-
-[dependencies]
-librefang-skill = "0.1"
-serde_json = "1"
-
-[profile.release]
-panic = "abort"
-"#
-                    ),
-                ),
-                (
-                    "src/lib.rs".to_string(),
-                    format!(
-                        r#"//! LibreFang skill: {name}
-use librefang_skill::{{skill, Request}};
-use serde_json::{{json, Value}};
-
-fn handle(req: Request) -> Result<Value, String> {{
-    match req.tool.as_str() {{
-        "{tool_name}" => {{
-            // TODO: Implement your skill logic here.
-            let input = req.input.get("input").and_then(Value::as_str).unwrap_or("");
-            Ok(json!({{ "result": format!("Processed: {{input}}") }}))
-        }}
-        other => Err(format!("unknown tool: {{other}}")),
-    }}
-}}
-
-skill!(handle);
-"#
-                    ),
-                ),
-            ],
-            build_steps: vec![
-                "rustup target add wasm32-unknown-unknown".to_string(),
-                "cargo build --release --target wasm32-unknown-unknown".to_string(),
-                "cp target/wasm32-unknown-unknown/release/skill.wasm skill.wasm".to_string(),
-            ],
-        },
-        other => {
-            eprintln!("Unsupported runtime '{other}'. Choose one of: python, node, wasm.");
-            std::process::exit(1);
-        }
-    };
-
-    let manifest = format!(
-        r#"[skill]
-name = "{name}"
-version = "{version}"
-description = "{description}"
-author = ""
-license = "MIT"
-tags = []
-
-[runtime]
-type = "{runtime}"
-entry = "{entry}"
-
-[[tools.provided]]
-name = "{tool_name}"
-description = "{description}"
-input_schema = {{ type = "object", properties = {{ input = {{ type = "string" }} }}, required = ["input"] }}
-
-[requirements]
-tools = []
-capabilities = []
-"#,
-        version = librefang_types::VERSION,
-        entry = scaffold.entry,
-    );
-
-    std::fs::write(skill_dir.join("skill.toml"), &manifest).unwrap();
-    for (rel, content) in &scaffold.files {
-        let path = skill_dir.join(rel);
-        if let Some(parent) = path.parent() {
-            std::fs::create_dir_all(parent).unwrap();
-        }
-        std::fs::write(&path, content).unwrap();
-    }
-
-    println!("\nSkill created: {}", skill_dir.display());
-    println!("\nFiles:");
-    println!("  skill.toml");
-    for (rel, _) in &scaffold.files {
-        println!("  {rel}");
-    }
-    println!("\nNext steps:");
-    let mut step = 1;
-    println!("  {step}. Edit the entry point to implement your skill logic");
-    for build_step in &scaffold.build_steps {
-        step += 1;
-        println!("  {step}. {build_step}");
-    }
-    step += 1;
-    println!(
-        "  {step}. Test locally: librefang skill test {}",
-        skill_dir.display()
-    );
-    step += 1;
-    println!(
-        "  {step}. Install: librefang skill install {}",
-        skill_dir.display()
-    );
-}
-
-/// Print an EvolutionResult as a one-line status.
-fn print_evolution_result(result: &librefang_skills::evolution::EvolutionResult) {
-    let marker = if result.success { "OK" } else { "FAIL" };
-    match &result.version {
-        Some(v) => println!("[{marker}] {} (v{v})", result.message),
-        None => println!("[{marker}] {}", result.message),
-    }
-}
-
-/// Resolve a skill by name. Respects `--hand` so evolve operations can
-/// target a per-hand workspace skills dir just like `install`/`list`.
-fn load_installed_skill(
-    name: &str,
-    hand: Option<&str>,
-) -> (PathBuf, librefang_skills::InstalledSkill) {
-    let skills_dir = resolve_skills_dir(hand);
-    let mut registry = librefang_skills::registry::SkillRegistry::new(skills_dir.clone());
-    if let Err(e) = registry.load_all() {
-        eprintln!("Error loading skill registry: {e}");
-        std::process::exit(1);
-    }
-    match registry.get(name) {
-        Some(skill) => (skills_dir, skill.clone()),
-        None => {
-            eprintln!("Skill '{name}' not found in {}", skills_dir.display());
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_skill_evolve(sub: EvolveCommands) {
-    match sub {
-        EvolveCommands::Create {
-            name,
-            description,
-            context_file,
-            tags,
-            hand,
-        } => {
-            let prompt_context = match read_file_or_stdin(&context_file) {
-                Ok(s) => s,
-                Err(e) => {
-                    eprintln!("Failed to read {}: {e}", context_file.display());
-                    std::process::exit(1);
-                }
-            };
-            let tag_list: Vec<String> = tags
-                .split(',')
-                .map(|t| t.trim())
-                .filter(|t| !t.is_empty())
-                .map(String::from)
-                .collect();
-            let skills_dir = resolve_skills_dir(hand.as_deref());
-            if let Err(e) = std::fs::create_dir_all(&skills_dir) {
-                eprintln!("Failed to create skills dir: {e}");
-                std::process::exit(1);
-            }
-            match librefang_skills::evolution::create_skill(
-                &skills_dir,
-                &name,
-                &description,
-                &prompt_context,
-                tag_list,
-                Some("cli"),
-            ) {
-                Ok(r) => print_evolution_result(&r),
-                Err(e) => {
-                    eprintln!("Create failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        EvolveCommands::Update {
-            name,
-            context_file,
-            changelog,
-            hand,
-        } => {
-            let new_ctx = match read_file_or_stdin(&context_file) {
-                Ok(s) => s,
-                Err(e) => {
-                    eprintln!("Failed to read {}: {e}", context_file.display());
-                    std::process::exit(1);
-                }
-            };
-            let (_, skill) = load_installed_skill(&name, hand.as_deref());
-            match librefang_skills::evolution::update_skill(
-                &skill,
-                &new_ctx,
-                &changelog,
-                Some("cli"),
-            ) {
-                Ok(r) => print_evolution_result(&r),
-                Err(e) => {
-                    eprintln!("Update failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        EvolveCommands::Patch {
-            name,
-            old_file,
-            new_file,
-            changelog,
-            replace_all,
-            hand,
-        } => {
-            let old_str = match read_file_or_stdin(&old_file) {
-                Ok(s) => s,
-                Err(e) => {
-                    eprintln!("Failed to read {}: {e}", old_file.display());
-                    std::process::exit(1);
-                }
-            };
-            let new_str = match read_file_or_stdin(&new_file) {
-                Ok(s) => s,
-                Err(e) => {
-                    eprintln!("Failed to read {}: {e}", new_file.display());
-                    std::process::exit(1);
-                }
-            };
-            let (_, skill) = load_installed_skill(&name, hand.as_deref());
-            match librefang_skills::evolution::patch_skill(
-                &skill,
-                &old_str,
-                &new_str,
-                &changelog,
-                replace_all,
-                Some("cli"),
-            ) {
-                Ok(r) => print_evolution_result(&r),
-                Err(e) => {
-                    eprintln!("Patch failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        EvolveCommands::Delete { name, hand } => {
-            let skills_dir = resolve_skills_dir(hand.as_deref());
-            match librefang_skills::evolution::delete_skill(&skills_dir, &name) {
-                Ok(r) => print_evolution_result(&r),
-                Err(e) => {
-                    eprintln!("Delete failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        EvolveCommands::Rollback { name, hand } => {
-            let (_, skill) = load_installed_skill(&name, hand.as_deref());
-            match librefang_skills::evolution::rollback_skill(&skill, Some("cli")) {
-                Ok(r) => print_evolution_result(&r),
-                Err(e) => {
-                    eprintln!("Rollback failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        EvolveCommands::WriteFile {
-            name,
-            path,
-            source,
-            hand,
-        } => {
-            let content = match read_file_or_stdin(&source) {
-                Ok(s) => s,
-                Err(e) => {
-                    eprintln!("Failed to read {}: {e}", source.display());
-                    std::process::exit(1);
-                }
-            };
-            let (_, skill) = load_installed_skill(&name, hand.as_deref());
-            match librefang_skills::evolution::write_supporting_file(&skill, &path, &content) {
-                Ok(r) => print_evolution_result(&r),
-                Err(e) => {
-                    eprintln!("Write-file failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        EvolveCommands::RemoveFile { name, path, hand } => {
-            let (_, skill) = load_installed_skill(&name, hand.as_deref());
-            match librefang_skills::evolution::remove_supporting_file(&skill, &path) {
-                Ok(r) => print_evolution_result(&r),
-                Err(e) => {
-                    eprintln!("Remove-file failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        EvolveCommands::History { name, json, hand } => {
-            let (_, skill) = load_installed_skill(&name, hand.as_deref());
-            let meta = librefang_skills::evolution::get_evolution_info(&skill);
-            if json {
-                match serde_json::to_string_pretty(&meta) {
-                    Ok(s) => println!("{s}"),
-                    Err(e) => {
-                        eprintln!("Failed to serialize history: {e}");
-                        std::process::exit(1);
-                    }
-                }
-                return;
-            }
-            println!("Skill: {}", skill.manifest.skill.name);
-            println!("Current version: {}", skill.manifest.skill.version);
-            println!("Use count: {}", meta.use_count);
-            println!("Evolution count: {}", meta.evolution_count);
-            if meta.versions.is_empty() {
-                println!("\nNo version history recorded.");
-                return;
-            }
-            println!();
-            let mut t = crate::table::Table::new(&["VERSION", "TIMESTAMP", "CHANGELOG"]);
-            for v in meta.versions.iter().rev() {
-                t.add_row(&[&v.version, &v.timestamp, &v.changelog]);
-            }
-            t.print();
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Skill workshop pending review (#3328)
-// ---------------------------------------------------------------------------
-
-fn cmd_skill_pending(sub: PendingCommands) {
-    let skills_root = librefang_home().join("skills");
-    match sub {
-        PendingCommands::List { agent } => {
-            let candidates = match &agent {
-                Some(a) => librefang_kernel::skill_workshop::storage::list_pending(&skills_root, a),
-                None => librefang_kernel::skill_workshop::storage::list_pending_all(&skills_root),
-            };
-            let candidates = match candidates {
-                Ok(v) => v,
-                Err(e) => {
-                    eprintln!("Failed to read pending directory: {e}");
-                    std::process::exit(1);
-                }
-            };
-            if candidates.is_empty() {
-                println!(
-                    "No pending skill candidates.{}",
-                    match &agent {
-                        Some(a) => format!(" (filter: agent {a})"),
-                        None => String::new(),
-                    }
-                );
-                return;
-            }
-            println!("{:<38}  {:<18}  {:<22}  NAME", "ID", "SOURCE", "CAPTURED");
-            for c in candidates {
-                let source_label = match &c.source {
-                    librefang_kernel::skill_workshop::CaptureSource::ExplicitInstruction {
-                        ..
-                    } => "explicit_instr",
-                    librefang_kernel::skill_workshop::CaptureSource::UserCorrection { .. } => {
-                        "user_correction"
-                    }
-                    librefang_kernel::skill_workshop::CaptureSource::RepeatedToolPattern {
-                        ..
-                    } => "tool_pattern",
-                };
-                println!(
-                    "{:<38}  {:<18}  {:<22}  {}",
-                    c.id,
-                    source_label,
-                    c.captured_at.format("%Y-%m-%d %H:%M:%S UTC"),
-                    c.name
-                );
-            }
-        }
-        PendingCommands::Show { id } => {
-            let candidate = match librefang_kernel::skill_workshop::storage::load_candidate(
-                &skills_root,
-                &id,
-            ) {
-                Ok(c) => c,
-                Err(e) => {
-                    eprintln!("Failed to load candidate: {e}");
-                    std::process::exit(1);
-                }
-            };
-            let toml_str = match toml::to_string_pretty(&candidate) {
-                Ok(s) => s,
-                Err(e) => {
-                    eprintln!("Failed to render candidate as TOML: {e}");
-                    std::process::exit(1);
-                }
-            };
-            print!("{toml_str}");
-        }
-        PendingCommands::Approve { id } => {
-            match librefang_kernel::skill_workshop::storage::approve_candidate(
-                &skills_root,
-                &skills_root,
-                &id,
-            ) {
-                Ok(result) => {
-                    println!(
-                        "Approved candidate {} → installed skill '{}' (v{}).",
-                        id,
-                        result.skill_name,
-                        result.version.unwrap_or_else(|| "?".to_string())
-                    );
-                }
-                Err(e) => {
-                    eprintln!("Approve failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-        PendingCommands::Reject { id } => {
-            match librefang_kernel::skill_workshop::storage::reject_candidate(&skills_root, &id) {
-                Ok(()) => println!("Rejected and removed candidate {id}."),
-                Err(e) => {
-                    eprintln!("Reject failed: {e}");
-                    std::process::exit(1);
-                }
-            }
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Channel commands
-// ---------------------------------------------------------------------------
-
-// maybe_write_channel_config / notify_daemon_restart removed — they
-// supported the interactive in-process channel onboarding flow whose
-// callers were dropped when channels moved to sidecars, leaving both
-// helpers orphaned.
-
-// ---------------------------------------------------------------------------
-// Hand commands
-// ---------------------------------------------------------------------------
-
-fn cmd_hand_install(path: &str) {
-    let base = require_daemon("hand install");
-    let dir = std::path::Path::new(path);
-    let toml_path = dir.join("HAND.toml");
-    let skill_path = dir.join("SKILL.md");
-
-    if !toml_path.exists() {
-        eprintln!(
-            "Error: No HAND.toml found in {}",
-            dir.canonicalize()
-                .unwrap_or_else(|_| dir.to_path_buf())
-                .display()
-        );
-        std::process::exit(1);
-    }
-
-    let toml_content = std::fs::read_to_string(&toml_path).unwrap_or_else(|e| {
-        eprintln!("Error reading {}: {e}", toml_path.display());
-        std::process::exit(1);
-    });
-    let skill_content = std::fs::read_to_string(&skill_path).unwrap_or_default();
-
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/hands/install"))
-            .json(&serde_json::json!({
-                "toml_content": toml_content,
-                "skill_content": skill_content,
-            }))
-            .send(),
-    );
-
-    if let Some(err) = body.get("error").and_then(|v| v.as_str()) {
-        eprintln!("Error: {err}");
-        std::process::exit(1);
-    }
-
-    println!(
-        "Installed hand: {} ({})",
-        body["name"].as_str().unwrap_or("?"),
-        body["id"].as_str().unwrap_or("?"),
-    );
-    println!(
-        "Use `librefang hand activate {}` to start it.",
-        body["id"].as_str().unwrap_or("?")
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Channel commands (sidecar-aware). Replace the pre-#5463 in-process
-// wizards: every channel now runs out-of-process, configuration goes
-// through the surviving daemon endpoints (GET /api/channels for the
-// list, GET /api/channels/registry + POST /api/channels/sidecar/{name}/
-// configure for setup, POST /api/channels/reload to apply, plus a local
-// `rm` that strips a [[sidecar_channels]] entry from config.toml).
-// ---------------------------------------------------------------------------
-
-fn cmd_channel_list() {
-    let base = require_daemon("channel list");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/channels")).send());
-    let items = body
-        .get("items")
-        .and_then(|v| v.as_array())
-        .cloned()
-        .unwrap_or_default();
-    if items.is_empty() {
-        println!("No channels configured.");
-        println!("Use `librefang channel setup` to add one.");
-        return;
-    }
-    let mut t = crate::table::Table::new(&["NAME", "KIND", "CONFIGURED", "TOKEN", "24H MSGS"]);
-    for ch in &items {
-        let name = ch.get("name").and_then(|v| v.as_str()).unwrap_or("?");
-        let kind = ch.get("category").and_then(|v| v.as_str()).unwrap_or("?");
-        let configured = ch
-            .get("configured")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let has_token = ch
-            .get("has_token")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let msgs = ch.get("msgs_24h").and_then(|v| v.as_u64()).unwrap_or(0);
-        t.add_row(&[
-            name,
-            kind,
-            if configured { "yes" } else { "no" },
-            if has_token { "yes" } else { "no" },
-            &msgs.to_string(),
-        ]);
-    }
-    t.print();
-}
-
-fn cmd_channel_reload() {
-    let base = require_daemon("channel reload");
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/channels/reload"))
-            .json(&serde_json::json!({}))
-            .send(),
-    );
-    let started = body
-        .get("started")
-        .and_then(|v| v.as_array())
-        .map(|a| a.len())
-        .unwrap_or(0);
-    println!("Channels reloaded ({started} sidecar(s) started).");
-}
-
-fn cmd_channel_setup(name: Option<&str>) {
-    let base = require_daemon("channel setup");
-    let client = daemon_client();
-    // `GET /api/channels` carries the full sidecar describe schema for
-    // every discoverable adapter on `fields[]`, so we don't need a
-    // separate /registry call for the picker — same list does both
-    // jobs.
-    let body = daemon_json(client.get(format!("{base}/api/channels")).send());
-    let all: Vec<serde_json::Value> = body
-        .get("items")
-        .and_then(|v| v.as_array())
-        .cloned()
-        .unwrap_or_default();
-
-    // Resolve the target row: explicit `<NAME>` argument, or interactive
-    // picker over unconfigured rows.
-    let target = match name {
-        Some(n) => all
-            .iter()
-            .find(|c| c.get("name").and_then(|v| v.as_str()) == Some(n))
-            .cloned(),
-        None => {
-            // Distinguish the two empty-picker cases so the operator
-            // knows which is which:
-            //  - `all.is_empty()`: daemon's `GET /api/channels` returned
-            //    nothing at all — both `sidecar_channel_rows` and
-            //    `sidecar_discovery_rows` are empty. That means there
-            //    are no `[[sidecar_channels]]` entries AND nothing in
-            //    the SIDECAR_CATALOG (the latter is normally only
-            //    empty if the SDK wasn't installed alongside the
-            //    daemon — fix is `pip install librefang-sdk`).
-            //  - all non-empty but `candidates.is_empty()`: the
-            //    operator has configured every adapter the catalog
-            //    knows about. Use `librefang channel list` to see /
-            //    `librefang channel rm <name>` to drop one.
-            if all.is_empty() {
-                println!("Daemon's channel registry is empty.");
-                println!("Install the sidecar SDK so adapters appear in the catalog:");
-                println!("  pip install librefang-sdk");
-                println!("Then re-run `librefang channel setup`.");
-                return;
-            }
-            let candidates: Vec<&serde_json::Value> = all
-                .iter()
-                .filter(|c| c.get("configured").and_then(|v| v.as_bool()) != Some(true))
-                .collect();
-            if candidates.is_empty() {
-                println!("Every available channel is already configured.");
-                println!("Use `librefang channel list` to see them, or");
-                println!("`librefang channel rm <name>` to remove an entry first.");
-                return;
-            }
-            println!("Pick a channel to set up:");
-            for (i, ch) in candidates.iter().enumerate() {
-                let n = ch.get("name").and_then(|v| v.as_str()).unwrap_or("?");
-                let d = ch.get("display_name").and_then(|v| v.as_str()).unwrap_or(n);
-                println!("  {:>2}. {:<14} {}", i + 1, n, d);
-            }
-            let choice = prompt_input("Choice [1]: ");
-            let idx = if choice.trim().is_empty() {
-                0
-            } else {
-                choice
-                    .trim()
-                    .parse::<usize>()
-                    .unwrap_or(1)
-                    .saturating_sub(1)
-                    .min(candidates.len() - 1)
-            };
-            Some(candidates[idx].clone())
-        }
-    };
-    let target = match target {
-        Some(t) => t,
-        None => {
-            ui::error_with_fix(
-                &format!("Unknown channel: {}", name.unwrap_or("?")),
-                "Run `librefang channel list` to see the available adapters.",
-            );
-            std::process::exit(1);
-        }
-    };
-    let chan_name = target
-        .get("name")
-        .and_then(|v| v.as_str())
-        .unwrap_or("")
-        .to_string();
-    let fields: Vec<serde_json::Value> = target
-        .get("fields")
-        .and_then(|v| v.as_array())
-        .cloned()
-        .unwrap_or_default();
-    if fields.is_empty() {
-        println!("`{chan_name}` exposes no configurable fields — nothing to prompt for.");
-        println!("(Hot-reload anyway with `librefang channel reload` if you've already edited config.toml by hand.)");
-        return;
-    }
-
-    let mut values = serde_json::Map::new();
-    for f in &fields {
-        let key = f.get("key").and_then(|v| v.as_str()).unwrap_or_default();
-        if key.is_empty() {
-            continue;
-        }
-        let label = f.get("label").and_then(|v| v.as_str()).unwrap_or(key);
-        let required = f.get("required").and_then(|v| v.as_bool()).unwrap_or(false);
-        let ftype = f.get("type").and_then(|v| v.as_str()).unwrap_or("text");
-        let has_value = f
-            .get("has_value")
-            .and_then(|v| v.as_bool())
-            .unwrap_or(false);
-        let current = f.get("value").and_then(|v| v.as_str()).unwrap_or("");
-
-        // Secret-typed + has_value=true: blank means "keep existing".
-        // Non-secret + has current value: show as default-in-brackets.
-        let prompt = if ftype == "secret" && has_value {
-            format!("  {label} ({key}) [set — leave blank to keep]: ")
-        } else if !current.is_empty() {
-            format!("  {label} ({key}) [{current}]: ")
-        } else if required {
-            format!("  {label} ({key}) *: ")
-        } else {
-            format!("  {label} ({key}): ")
-        };
-        let entered = prompt_input(&prompt);
-        let val = entered.trim();
-        if val.is_empty() {
-            continue;
-        }
-        values.insert(key.to_string(), serde_json::Value::String(val.to_string()));
-    }
-
-    // Sidecar names come from `SIDECAR_CATALOG` keys — short
-    // alphanumeric (`telegram`, `ntfy`, …), URL-safe as-is. No need
-    // for percent-encoding.
-    let url = format!("{base}/api/channels/sidecar/{chan_name}/configure");
-    let payload = serde_json::json!({ "values": values });
-    let body = daemon_json(client.post(&url).json(&payload).send());
-    // `daemon_json` only logs 5xx; 4xx silently returns the error body.
-    // Surface those by checking for the SidecarSaveResult shape. The
-    // `ApiErrorResponse` envelope (see librefang-api types.rs:114-164)
-    // serializes the human-readable message at both `error.message`
-    // (nested, #3639 preferred shape) and `message` (top-level flat
-    // alias kept for legacy callers); prefer the nested one, fall
-    // through to the flat alias for older deployments.
-    if body.get("status").and_then(|v| v.as_str()) != Some("saved") {
-        let err = body
-            .pointer("/error/message")
-            .and_then(|v| v.as_str())
-            .or_else(|| body.get("message").and_then(|v| v.as_str()))
-            .unwrap_or("save failed (no error body)");
-        ui::error_with_fix(
-            &format!("Save for `{chan_name}` rejected: {err}"),
-            "Re-run with corrected values, or check the daemon log for details.",
-        );
-        std::process::exit(1);
-    }
-    let restart_required = body
-        .get("restart_required")
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-    let shadowed = body
-        .get("shadowed_secrets")
-        .and_then(|v| v.as_array())
-        .cloned()
-        .unwrap_or_default();
-    if restart_required {
-        println!("✓ Saved `{chan_name}` — restart the daemon for changes to apply.");
-    } else {
-        println!("✓ Saved `{chan_name}` — hot-reload applied.");
-    }
-    if !shadowed.is_empty() {
-        let keys: Vec<String> = shadowed
-            .iter()
-            .filter_map(|v| v.as_str().map(String::from))
-            .collect();
-        eprintln!(
-            "Warning: shell environment variables shadow these tokens — unset them and restart for the new value to take effect: {}",
-            keys.join(", "),
-        );
-    }
-}
-
-fn cmd_channel_rm(name: &str) {
-    // Strip the matching `[[sidecar_channels]]` entry from
-    // ~/.librefang/config.toml in-place, then trigger a daemon reload
-    // (best-effort: if no daemon is running, the file edit is enough
-    // — the next daemon start will pick up the changed config).
-    let home = cli_librefang_home();
-    let path = home.join("config.toml");
-    let original = match std::fs::read_to_string(&path) {
-        Ok(s) => s,
-        Err(e) => {
-            ui::error_with_fix(
-                &format!("Cannot read {}: {e}", path.display()),
-                "Run `librefang init` to create the config file.",
-            );
-            std::process::exit(1);
-        }
-    };
-    let mut doc: toml_edit::DocumentMut = match original.parse() {
-        Ok(d) => d,
-        Err(e) => {
-            ui::error_with_fix(
-                &format!("Cannot parse {}: {e}", path.display()),
-                "Fix the TOML syntax and retry.",
-            );
-            std::process::exit(1);
-        }
-    };
-    let arr = match doc
-        .get_mut("sidecar_channels")
-        .and_then(|v| v.as_array_of_tables_mut())
-    {
-        Some(a) => a,
-        None => {
-            println!("No [[sidecar_channels]] entries in config.toml — nothing to remove.");
-            return;
-        }
-    };
-    // `toml_edit::ArrayOfTables` has no `retain`; collect matching indices
-    // then remove in reverse so earlier indices stay stable.
-    let to_remove: Vec<usize> = arr
-        .iter()
-        .enumerate()
-        .filter_map(|(i, t)| match t.get("name").and_then(|v| v.as_str()) {
-            Some(n) if n == name => Some(i),
-            _ => None,
-        })
-        .collect();
-    let removed = to_remove.len();
-    for &i in to_remove.iter().rev() {
-        arr.remove(i);
-    }
-    if removed == 0 {
-        println!("No [[sidecar_channels]] entry with name=\"{name}\".");
-        return;
-    }
-    if let Err(e) = std::fs::write(&path, doc.to_string()) {
-        ui::error_with_fix(
-            &format!("Failed to write {}: {e}", path.display()),
-            "Check filesystem permissions.",
-        );
-        std::process::exit(1);
-    }
-    println!("✓ Removed {removed} [[sidecar_channels]] entry/entries named `{name}`.");
-    match find_daemon() {
-        Some(base) => {
-            let client = daemon_client();
-            match client
-                .post(format!("{base}/api/channels/reload"))
-                .json(&serde_json::json!({}))
-                .send()
-            {
-                Ok(r) if r.status().is_success() => println!("  Hot-reloaded daemon."),
-                Ok(r) => eprintln!(
-                    "  Reload returned {}: change will apply on next daemon restart.",
-                    r.status()
-                ),
-                Err(e) => eprintln!(
-                    "  Could not contact daemon for reload ({e}); change will apply on next start."
-                ),
-            }
-        }
-        None => println!("  Daemon not running; change will apply on next start."),
-    }
-}
-
-fn cmd_hand_list() {
-    let base = require_daemon("hand list");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/hands")).send());
-    // API returns {"hands": [...]} or a bare array
-    let arr_val;
-    if let Some(arr) = body.get("hands").and_then(|v| v.as_array()) {
-        arr_val = arr.clone();
-    } else if let Some(arr) = body.as_array() {
-        arr_val = arr.clone();
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-        return;
-    }
-    if let Some(arr) = Some(&arr_val) {
-        if arr.is_empty() {
-            println!("No hands available.");
-            return;
-        }
-        let mut t = crate::table::Table::new(&["ID", "NAME", "CATEGORY", "DESCRIPTION"]);
-        for h in arr {
-            t.add_row(&[
-                h["id"].as_str().unwrap_or("?"),
-                h["name"].as_str().unwrap_or("?"),
-                h["category"].as_str().unwrap_or("?"),
-                &h["description"]
-                    .as_str()
-                    .unwrap_or("")
-                    .chars()
-                    .take(40)
-                    .collect::<String>(),
-            ]);
-        }
-        t.print();
-        println!("\nUse `librefang hand activate <id>` to activate a hand.");
-    }
-}
-
-fn cmd_hand_active() {
-    let base = require_daemon("hand active");
-    let client = daemon_client();
-    let arr = fetch_active_hand_instances(&base, &client);
-    if arr.is_empty() {
-        println!("No active hands.");
-        return;
-    }
-    let mut t = crate::table::Table::new(&["INSTANCE", "HAND", "STATUS", "AGENT"]);
-    for i in &arr {
-        t.add_row(&[
-            i["instance_id"].as_str().unwrap_or("?"),
-            i["hand_id"].as_str().unwrap_or("?"),
-            i["status"].as_str().unwrap_or("?"),
-            i["agent_name"].as_str().unwrap_or("?"),
-        ]);
-    }
-    t.print();
-}
-
-fn cmd_hand_status(id: Option<&str>) {
-    if id.is_none() {
-        cmd_hand_active();
-        return;
-    }
-
-    let id = id.unwrap_or_default();
-    let base = require_daemon("hand status");
-    let client = daemon_client();
-    let active = fetch_active_hand_instances(&base, &client);
-
-    if let Some(instance) = resolve_hand_instance(&active, id) {
-        let hand_id = instance["hand_id"].as_str().unwrap_or(id);
-        let hand_body = daemon_json(client.get(format!("{base}/api/hands/{hand_id}")).send());
-        let name = hand_body["name"].as_str().unwrap_or(hand_id);
-        let status = instance["status"].as_str().unwrap_or("unknown");
-        let instance_id = instance["instance_id"].as_str().unwrap_or("?");
-        let agent_name = instance["agent_name"].as_str().unwrap_or("?");
-
-        ui::section("Hand Status");
-        ui::kv("Hand", hand_id);
-        ui::kv("Name", name);
-        ui::kv("Instance", instance_id);
-        ui::kv("Status", status);
-        ui::kv("Agent", agent_name);
-        return;
-    }
-
-    let hand_body = daemon_json(client.get(format!("{base}/api/hands/{id}")).send());
-    if hand_body.get("error").is_some() {
-        ui::error(&format!(
-            "No active hand or installed hand found for '{id}'."
-        ));
-        std::process::exit(1);
-    }
-
-    ui::section("Hand Status");
-    ui::kv("Hand", hand_body["id"].as_str().unwrap_or(id));
-    ui::kv("Name", hand_body["name"].as_str().unwrap_or(id));
-    ui::kv("Status", "inactive");
-    if let Some(description) = hand_body["description"].as_str() {
-        if !description.is_empty() {
-            ui::kv("Description", description);
-        }
-    }
-}
-
-fn cmd_hand_activate(id: &str) {
-    let base = require_daemon("hand activate");
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/hands/{id}/activate"))
-            .header("content-type", "application/json")
-            .body("{}")
-            .send(),
-    );
-    if body.get("instance_id").is_some() {
-        println!(
-            "Hand '{}' activated (instance: {}, agent: {})",
-            id,
-            body["instance_id"].as_str().unwrap_or("?"),
-            body["agent_name"].as_str().unwrap_or("?"),
-        );
-    } else {
-        eprintln!(
-            "Failed to activate hand '{}': {}",
-            id,
-            body["error"].as_str().unwrap_or("Unknown error")
-        );
-        std::process::exit(1);
-    }
-}
-
-fn cmd_hand_deactivate(id: &str) {
-    let base = require_daemon("hand deactivate");
-    let client = daemon_client();
-    // First find the instance ID for this hand
-    let arr = fetch_active_hand_instances(&base, &client);
-    let instance_id = arr.iter().find_map(|i| {
-        if i["hand_id"].as_str() == Some(id) {
-            i["instance_id"].as_str().map(|s| s.to_string())
-        } else {
-            None
-        }
-    });
-
-    match instance_id {
-        Some(iid) => {
-            let body = daemon_json(
-                client
-                    .delete(format!("{base}/api/hands/instances/{iid}"))
-                    .send(),
-            );
-            if body.get("status").is_some() {
-                println!("Hand '{id}' deactivated.");
-            } else {
-                eprintln!(
-                    "Failed: {}",
-                    body["error"].as_str().unwrap_or("Unknown error")
-                );
-                std::process::exit(1);
-            }
-        }
-        None => {
-            eprintln!("No active instance found for hand '{id}'.");
-            std::process::exit(1);
-        }
-    }
-}
-
-fn cmd_hand_info(id: &str) {
-    let base = require_daemon("hand info");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/hands/{id}")).send());
-    if body.get("error").is_some() {
-        eprintln!("Hand not found: {}", body["error"].as_str().unwrap_or(id));
-        std::process::exit(1);
-    }
-    println!(
-        "{}",
-        serde_json::to_string_pretty(&body).unwrap_or_default()
-    );
-}
-
-fn cmd_hand_check_deps(id: &str) {
-    let base = require_daemon("hand check-deps");
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/hands/{id}/check-deps"))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&format!(
-            "Failed: {}",
-            body["error"].as_str().unwrap_or("?")
-        ));
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
-fn cmd_hand_install_deps(id: &str) {
-    let base = require_daemon("hand install-deps");
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/hands/{id}/install-deps"))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&format!(
-            "Failed: {}",
-            body["error"].as_str().unwrap_or("?")
-        ));
-    } else {
-        ui::success(&i18n::t_args("hand-install-deps-success", &[("id", id)]));
-        if let Some(results) = body.get("results") {
-            println!(
-                "{}",
-                serde_json::to_string_pretty(results).unwrap_or_default()
-            );
-        }
-    }
-}
-
-fn cmd_hand_pause(id: &str) {
-    let base = require_daemon("hand pause");
-    let client = daemon_client();
-    let active = fetch_active_hand_instances(&base, &client);
-    let resolved = resolve_hand_instance(&active, id);
-    let instance_id = resolved
-        .as_ref()
-        .and_then(|instance| instance["instance_id"].as_str())
-        .unwrap_or(id);
-    let hand_label = resolved
-        .as_ref()
-        .and_then(|instance| instance["hand_id"].as_str())
-        .unwrap_or(id);
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/hands/instances/{instance_id}/pause"))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&format!(
-            "Failed: {}",
-            body["error"].as_str().unwrap_or("?")
-        ));
-        std::process::exit(1);
-    } else {
-        ui::success(&i18n::t_args(
-            "hand-paused",
-            &[("id", &format!("{hand_label} (instance: {instance_id})"))],
-        ));
-    }
-}
-
-fn cmd_hand_resume(id: &str) {
-    let base = require_daemon("hand resume");
-    let client = daemon_client();
-    let active = fetch_active_hand_instances(&base, &client);
-    let resolved = resolve_hand_instance(&active, id);
-    let instance_id = resolved
-        .as_ref()
-        .and_then(|instance| instance["instance_id"].as_str())
-        .unwrap_or(id);
-    let hand_label = resolved
-        .as_ref()
-        .and_then(|instance| instance["hand_id"].as_str())
-        .unwrap_or(id);
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/hands/instances/{instance_id}/resume"))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&format!(
-            "Failed: {}",
-            body["error"].as_str().unwrap_or("?")
-        ));
-        std::process::exit(1);
-    } else {
-        ui::success(&i18n::t_args(
-            "hand-resumed",
-            &[("id", &format!("{hand_label} (instance: {instance_id})"))],
-        ));
-    }
-}
-
-fn cmd_hand_settings(id: &str) {
-    let base = require_daemon("hand settings");
-    let client = daemon_client();
-    let body = daemon_json(client.get(format!("{base}/api/hands/{id}/settings")).send());
-    if body.get("error").is_some() {
-        ui::error(&format!(
-            "Failed: {}",
-            body["error"].as_str().unwrap_or("?")
-        ));
-        std::process::exit(1);
-    }
-    if let Some(config) = body.get("config").and_then(|c| c.as_object()) {
-        if config.is_empty() {
-            ui::step(&format!("Hand '{id}' has no configurable settings."));
-        } else {
-            ui::section(&format!("Settings for '{id}'"));
-            for (k, v) in config {
-                println!("  {}: {}", k.bold(), v);
-            }
-        }
-    } else {
-        ui::step(&format!("Hand '{id}' has no configurable settings."));
-    }
-}
-
-fn cmd_hand_set(id: &str, key: &str, value: &str) {
-    let base = require_daemon("hand set");
-    let client = daemon_client();
-    let mut config = serde_json::Map::new();
-    config.insert(
-        key.to_string(),
-        serde_json::Value::String(value.to_string()),
-    );
-    let body = daemon_json(
-        client
-            .put(format!("{base}/api/hands/{id}/settings"))
-            .json(&serde_json::json!({ "config": config }))
-            .send(),
-    );
-    if body.get("error").is_some() {
-        ui::error(&format!(
-            "Failed: {}",
-            body["error"].as_str().unwrap_or("?")
-        ));
-        std::process::exit(1);
-    }
-    ui::success(&format!("Set {key}={value} for hand '{id}'."));
-}
-
-fn cmd_hand_reload() {
-    let base = require_daemon("hand reload");
-    let client = daemon_client();
-    let body = daemon_json(client.post(format!("{base}/api/hands/reload")).send());
-    if body.get("error").is_some() {
-        ui::error(&format!(
-            "Failed: {}",
-            body["error"].as_str().unwrap_or("?")
-        ));
-        std::process::exit(1);
-    }
-    let added = body["added"].as_u64().unwrap_or(0);
-    let updated = body["updated"].as_u64().unwrap_or(0);
-    let total = body["total"].as_u64().unwrap_or(0);
-    ui::success(&format!(
-        "Reloaded hands: {added} added, {updated} updated, {total} total."
-    ));
-}
-
-fn cmd_hand_chat(id: &str) {
-    let base = require_daemon("hand chat");
-    let client = daemon_client();
-    let active = fetch_active_hand_instances(&base, &client);
-    let resolved = match resolve_hand_instance(&active, id) {
-        Some(instance) => instance,
-        None => {
-            ui::error(&format!("No active hand instance found for '{id}'."));
-            ui::hint("Activate it first: librefang hand activate");
-            std::process::exit(1);
-        }
-    };
-    let instance_id = resolved["instance_id"]
-        .as_str()
-        .expect("instance_id missing");
-    let hand_id = resolved["hand_id"].as_str().unwrap_or(id);
-    let hand_name = resolved["hand_name"]
-        .as_str()
-        .or_else(|| resolved["name"].as_str())
-        .unwrap_or(hand_id);
-
-    install_ctrlc_handler();
-
-    println!(
-        "{} {} {}",
-        "Chat with".bold(),
-        hand_name.cyan().bold(),
-        "(type /quit to exit)".dimmed()
-    );
-    println!();
-
-    loop {
-        print!("{} ", "you >".green().bold());
-        io::stdout().flush().unwrap();
-        let mut line = String::new();
-        if io::stdin().lock().read_line(&mut line).unwrap_or(0) == 0 {
-            break; // EOF
-        }
-        let msg = line.trim();
-        if msg.is_empty() {
-            continue;
-        }
-        if msg == "/quit" || msg == "/exit" || msg == "/q" {
-            break;
-        }
-
-        let resp = client
-            .post(format!("{base}/api/hands/instances/{instance_id}/message"))
-            .json(&serde_json::json!({"message": msg}))
-            .send();
-
-        let body = daemon_json(resp);
-        if let Some(err) = body["error"].as_str() {
-            ui::error(err);
-            continue;
-        }
-        let reply = body["response"]
-            .as_str()
-            .or_else(|| body["reply"].as_str())
-            .unwrap_or("[no response]");
-        println!("{} {}\n", format!("{hand_name} >").cyan().bold(), reply);
-    }
-}
-
-fn fetch_active_hand_instances(
-    base: &str,
-    client: &reqwest::blocking::Client,
-) -> Vec<serde_json::Value> {
-    let body = daemon_json(client.get(format!("{base}/api/hands/active")).send());
-    body.get("instances")
-        .and_then(|v| v.as_array())
-        .or_else(|| body.as_array())
-        .cloned()
-        .unwrap_or_default()
-}
-
-fn resolve_hand_instance(
-    active_instances: &[serde_json::Value],
-    id_or_hand: &str,
-) -> Option<serde_json::Value> {
-    active_instances
-        .iter()
-        .find(|instance| {
-            instance["instance_id"].as_str() == Some(id_or_hand)
-                || instance["hand_id"].as_str() == Some(id_or_hand)
-        })
-        .cloned()
-}
-
-// ---------------------------------------------------------------------------
 // Config commands
 // ---------------------------------------------------------------------------
 
@@ -5191,393 +2675,6 @@ fn cmd_config_test_key(provider: &str) {
         ));
         std::process::exit(1);
     }
-}
-
-// ---------------------------------------------------------------------------
-// Quick chat (OpenClaw alias)
-// ---------------------------------------------------------------------------
-
-fn cmd_quick_chat(config: Option<PathBuf>, agent: Option<String>) {
-    ensure_initialized(&config);
-    tui::chat_runner::run_chat_tui(config, agent);
-}
-
-// ---------------------------------------------------------------------------
-// MCP server commands (librefang mcp {add,remove,list,catalog})
-// ---------------------------------------------------------------------------
-
-fn cmd_mcp_add(name: &str, key: Option<&str>) {
-    let home = librefang_home();
-    let mut catalog = librefang_extensions::catalog::McpCatalog::new(&home);
-    catalog.load(&home);
-
-    // Check template exists
-    let template = match catalog.get(name) {
-        Some(t) => t.clone(),
-        None => {
-            ui::error(&format!("Unknown MCP catalog entry: '{name}'"));
-            println!("\nAvailable MCP servers (catalog):");
-            for t in catalog.list() {
-                println!("  {} {} — {}", t.icon, t.id, t.description);
-            }
-            std::process::exit(1);
-        }
-    };
-
-    // Reject re-install of an already-configured server by name/template_id.
-    // The API path returns 409 here; the CLI was silently overwriting the
-    // existing [[mcp_servers]] entry (including edited transport/env/oauth)
-    // because upsert_mcp_server_local replaces by name. Users should remove
-    // first if they want to re-install.
-    let config_path = home.join("config.toml");
-    if config_path.is_file() {
-        let content = match std::fs::read_to_string(&config_path) {
-            Ok(c) => c,
-            Err(e) => {
-                ui::error(&format!("Failed to read {}: {e}", config_path.display()));
-                std::process::exit(1);
-            }
-        };
-        let parsed: toml::value::Table = match toml::from_str(&content) {
-            Ok(t) => t,
-            Err(e) => {
-                ui::error(&format!("{} is not valid TOML: {e}", config_path.display()));
-                std::process::exit(1);
-            }
-        };
-        if let Some(toml::Value::Array(servers)) = parsed.get("mcp_servers") {
-            let conflict = servers.iter().any(|v| {
-                let t = match v.as_table() {
-                    Some(t) => t,
-                    None => return false,
-                };
-                let matches_field = |k: &str| t.get(k).and_then(|n| n.as_str()) == Some(name);
-                matches_field("name") || matches_field("template_id")
-            });
-            if conflict {
-                ui::error(&format!(
-                    "MCP server '{name}' is already configured. Run \
-                     `librefang mcp remove {name}` first if you want to re-install."
-                ));
-                std::process::exit(1);
-            }
-        }
-    }
-
-    // Set up credential resolver (vault + dotenv + interactive prompt fallback)
-    let dotenv_path = home.join(".env");
-    let vault_path = home.join("vault.enc");
-    let vault = if vault_path.exists() {
-        let mut v = librefang_extensions::vault::CredentialVault::new(vault_path);
-        if v.unlock().is_ok() {
-            Some(v)
-        } else {
-            None
-        }
-    } else {
-        None
-    };
-    let mut resolver =
-        librefang_extensions::credentials::CredentialResolver::new(vault, Some(&dotenv_path))
-            .with_interactive(true);
-
-    // Build provided keys map
-    let mut provided_keys = std::collections::HashMap::new();
-    if let Some(key_value) = key {
-        // Auto-detect which env var to use (first required_env that's a secret)
-        if let Some(env_var) = template.required_env.iter().find(|e| e.is_secret) {
-            provided_keys.insert(env_var.name.clone(), key_value.to_string());
-        }
-    }
-
-    let result = match librefang_extensions::installer::install_integration(
-        &catalog,
-        &mut resolver,
-        name,
-        &provided_keys,
-    ) {
-        Ok(r) => r,
-        Err(e) => {
-            ui::error(&e.to_string());
-            std::process::exit(1);
-        }
-    };
-
-    // Persist the new [[mcp_servers]] entry directly into config.toml.
-    let config_path = home.join("config.toml");
-    if let Err(e) = upsert_mcp_server_local(&config_path, &result.server) {
-        ui::error(&format!("Failed to write config.toml: {e}"));
-        std::process::exit(1);
-    }
-
-    match &result.status {
-        librefang_types::mcp::McpStatus::Ready => ui::success(&result.message),
-        librefang_types::mcp::McpStatus::Setup => {
-            println!("{}", result.message.yellow());
-            println!("\nTo add credentials:");
-            for env in &template.required_env {
-                if env.is_secret {
-                    println!("  librefang vault set {}  # {}", env.name, env.help);
-                    if let Some(ref url) = env.get_url {
-                        println!("  Get it here: {url}");
-                    }
-                }
-            }
-        }
-        _ => println!("{}", result.message),
-    }
-
-    // If daemon is running, trigger hot-reload.
-    if let Some(base_url) = find_daemon() {
-        let client = daemon_client();
-        let _ = client.post(format!("{base_url}/api/mcp/reload")).send();
-    }
-}
-
-fn cmd_mcp_remove(name: &str) {
-    let home = librefang_home();
-    let config_path = home.join("config.toml");
-
-    // Resolve by template_id first, fall back to server name.
-    let target_name: Option<String> = {
-        let raw = std::fs::read_to_string(&config_path).unwrap_or_default();
-        let doc: toml::Value =
-            toml::from_str(&raw).unwrap_or(toml::Value::Table(Default::default()));
-        doc.as_table()
-            .and_then(|t| t.get("mcp_servers"))
-            .and_then(|v| v.as_array())
-            .and_then(|arr| {
-                arr.iter().find_map(|entry| {
-                    let tbl = entry.as_table()?;
-                    let tid = tbl.get("template_id").and_then(|v| v.as_str());
-                    let nm = tbl.get("name").and_then(|v| v.as_str())?;
-                    if tid == Some(name) || nm == name {
-                        Some(nm.to_string())
-                    } else {
-                        None
-                    }
-                })
-            })
-    };
-
-    let target_name = match target_name {
-        Some(n) => n,
-        None => {
-            ui::error(&format!("MCP server '{name}' is not configured"));
-            std::process::exit(1);
-        }
-    };
-
-    if let Err(e) = remove_mcp_server_local(&config_path, &target_name) {
-        ui::error(&format!("Failed to update config.toml: {e}"));
-        std::process::exit(1);
-    }
-
-    ui::success(&format!("{target_name} removed."));
-
-    // Hot-reload daemon
-    if let Some(base_url) = find_daemon() {
-        let client = daemon_client();
-        let _ = client.post(format!("{base_url}/api/mcp/reload")).send();
-    }
-}
-
-fn cmd_mcp_catalog(query: Option<&str>) {
-    let home = librefang_home();
-    let mut catalog = librefang_extensions::catalog::McpCatalog::new(&home);
-    catalog.load(&home);
-
-    // Installed state comes from config.mcp_servers' template_id field.
-    let installed_template_ids: std::collections::HashSet<String> = {
-        let raw = std::fs::read_to_string(home.join("config.toml")).unwrap_or_default();
-        toml::from_str::<toml::Value>(&raw)
-            .ok()
-            .and_then(|v| v.as_table().cloned())
-            .and_then(|t| t.get("mcp_servers").cloned())
-            .and_then(|v| v.as_array().cloned())
-            .map(|arr| {
-                arr.into_iter()
-                    .filter_map(|v| {
-                        v.as_table()
-                            .and_then(|t| t.get("template_id"))
-                            .and_then(|t| t.as_str())
-                            .map(|s| s.to_string())
-                    })
-                    .collect()
-            })
-            .unwrap_or_default()
-    };
-
-    let entries: Vec<_> = if let Some(q) = query {
-        catalog.search(q).into_iter().cloned().collect()
-    } else {
-        catalog.list().into_iter().cloned().collect()
-    };
-
-    if entries.is_empty() {
-        if let Some(q) = query {
-            println!("No MCP catalog entries matching '{q}'.");
-        } else {
-            println!("No MCP catalog entries available.");
-        }
-        return;
-    }
-
-    // Group by category
-    let mut by_category: std::collections::BTreeMap<
-        String,
-        Vec<&librefang_types::mcp::McpCatalogEntry>,
-    > = std::collections::BTreeMap::new();
-    for entry in &entries {
-        by_category
-            .entry(entry.category.to_string())
-            .or_default()
-            .push(entry);
-    }
-
-    for (category, items) in &by_category {
-        println!("\n{}", format!("  {category}").bold());
-        for item in items {
-            let status_badge = if installed_template_ids.contains(&item.id) {
-                "[Installed]".green().to_string()
-            } else {
-                "[Available]".dimmed().to_string()
-            };
-            println!(
-                "    {} {:<20} {:<13} {}",
-                item.icon, item.id, status_badge, item.description
-            );
-        }
-    }
-    println!();
-    println!(
-        "  {} catalog entries ({} installed)",
-        entries.len(),
-        entries
-            .iter()
-            .filter(|e| installed_template_ids.contains(&e.id))
-            .count()
-    );
-    println!("  Use `librefang mcp add <id>` to install an MCP server.");
-}
-
-fn cmd_mcp_list() {
-    let home = librefang_home();
-    let raw = std::fs::read_to_string(home.join("config.toml")).unwrap_or_default();
-    let doc: toml::Value = toml::from_str(&raw).unwrap_or(toml::Value::Table(Default::default()));
-    let servers = doc
-        .as_table()
-        .and_then(|t| t.get("mcp_servers"))
-        .and_then(|v| v.as_array());
-    let Some(servers) = servers else {
-        println!("No MCP servers configured.");
-        return;
-    };
-    if servers.is_empty() {
-        println!("No MCP servers configured.");
-        return;
-    }
-    println!();
-    println!(
-        "  {:<28} {:<14} {:<18} details",
-        "name", "template_id", "transport"
-    );
-    for entry in servers {
-        let Some(tbl) = entry.as_table() else {
-            continue;
-        };
-        let name = tbl.get("name").and_then(|v| v.as_str()).unwrap_or("?");
-        let tid = tbl
-            .get("template_id")
-            .and_then(|v| v.as_str())
-            .unwrap_or("-");
-        let (transport, detail) = match tbl.get("transport").and_then(|v| v.as_table()) {
-            Some(t) => {
-                let ttype = t.get("type").and_then(|v| v.as_str()).unwrap_or("?");
-                let detail = match ttype {
-                    "stdio" => t
-                        .get("command")
-                        .and_then(|v| v.as_str())
-                        .unwrap_or("")
-                        .to_string(),
-                    "sse" | "http" => t
-                        .get("url")
-                        .and_then(|v| v.as_str())
-                        .unwrap_or("")
-                        .to_string(),
-                    _ => String::new(),
-                };
-                (ttype.to_string(), detail)
-            }
-            None => ("-".to_string(), String::new()),
-        };
-        println!("  {name:<28} {tid:<14} {transport:<18} {detail}");
-    }
-    println!();
-    println!("  Use `librefang mcp catalog` to list installable entries.");
-}
-
-/// Local upsert helper — mirrors the API's `upsert_mcp_server_config`.
-fn upsert_mcp_server_local(
-    config_path: &std::path::Path,
-    entry: &librefang_types::config::McpServerConfigEntry,
-) -> Result<(), String> {
-    let mut table: toml::value::Table = if config_path.exists() {
-        let content = std::fs::read_to_string(config_path).map_err(|e| e.to_string())?;
-        // Propagate parse errors instead of silently defaulting. A
-        // malformed config.toml would otherwise be overwritten as a new
-        // near-empty file, wiping unrelated sections the user may want
-        // to fix by hand.
-        toml::from_str(&content).map_err(|e| format!("config.toml is not valid TOML: {e}"))?
-    } else {
-        toml::value::Table::new()
-    };
-
-    let entry_json = serde_json::to_value(entry).map_err(|e| e.to_string())?;
-    let entry_toml = json_to_toml_value_cli(&entry_json);
-
-    let servers = table
-        .entry("mcp_servers".to_string())
-        .or_insert_with(|| toml::Value::Array(Vec::new()));
-
-    if let toml::Value::Array(ref mut arr) = servers {
-        arr.retain(|v| {
-            v.as_table()
-                .and_then(|t| t.get("name"))
-                .and_then(|n| n.as_str())
-                .map(|n| n != entry.name)
-                .unwrap_or(true)
-        });
-        arr.push(entry_toml);
-    }
-
-    let toml_string = toml::to_string_pretty(&table).map_err(|e| e.to_string())?;
-    if let Some(parent) = config_path.parent() {
-        std::fs::create_dir_all(parent).map_err(|e| e.to_string())?;
-    }
-    std::fs::write(config_path, toml_string).map_err(|e| e.to_string())?;
-    Ok(())
-}
-
-fn remove_mcp_server_local(config_path: &std::path::Path, name: &str) -> Result<(), String> {
-    let mut table: toml::value::Table = if config_path.exists() {
-        let content = std::fs::read_to_string(config_path).map_err(|e| e.to_string())?;
-        toml::from_str(&content).map_err(|e| format!("config.toml is not valid TOML: {e}"))?
-    } else {
-        return Ok(());
-    };
-    if let Some(toml::Value::Array(ref mut arr)) = table.get_mut("mcp_servers") {
-        arr.retain(|v| {
-            v.as_table()
-                .and_then(|t| t.get("name"))
-                .and_then(|n| n.as_str())
-                .map(|n| n != name)
-                .unwrap_or(true)
-        });
-    }
-    let toml_string = toml::to_string_pretty(&table).map_err(|e| e.to_string())?;
-    std::fs::write(config_path, toml_string).map_err(|e| e.to_string())?;
-    Ok(())
 }
 
 // ---------------------------------------------------------------------------
@@ -7043,133 +4140,6 @@ fn cmd_cron_toggle(id: &str, enable: bool) {
     }
 }
 
-fn cmd_sessions(agent: Option<&str>, json: bool, active_only: bool) {
-    let base = require_daemon("sessions");
-    let client = daemon_client();
-    let url = match agent {
-        Some(a) => format!("{base}/api/sessions?agent={a}"),
-        None => format!("{base}/api/sessions"),
-    };
-    let body = daemon_json(client.get(&url).send());
-
-    // Build a (agent_id -> set<session_id>) map of currently-running sessions.
-    // Walks the unique agent ids in the listing once and asks the per-agent
-    // runtime endpoint added in #3172. Cheap on dev-scale agent counts; if
-    // this ever becomes a hotspot we can add a single-call /api/runtime.
-    let session_arr_owned: Option<Vec<serde_json::Value>> = body
-        .get("sessions")
-        .and_then(|v| v.as_array())
-        .cloned()
-        .or_else(|| body.as_array().cloned());
-    let mut active_sessions: std::collections::HashMap<String, std::collections::HashSet<String>> =
-        std::collections::HashMap::new();
-    if let Some(arr) = session_arr_owned.as_ref() {
-        let agent_ids: std::collections::HashSet<String> = arr
-            .iter()
-            .filter_map(|s| s["agent_id"].as_str().map(|id| id.to_string()))
-            .collect();
-        for aid in agent_ids {
-            let runtime_url = format!("{base}/api/agents/{aid}/runtime");
-            if let Ok(resp) = client.get(&runtime_url).send() {
-                if let Ok(items) = resp.json::<Vec<serde_json::Value>>() {
-                    let sids: std::collections::HashSet<String> = items
-                        .iter()
-                        .filter_map(|v| v["session_id"].as_str().map(|s| s.to_string()))
-                        .collect();
-                    active_sessions.insert(aid, sids);
-                }
-            }
-        }
-    }
-
-    let is_running = |s: &serde_json::Value| -> bool {
-        let aid = match s["agent_id"].as_str() {
-            Some(a) => a,
-            None => return false,
-        };
-        let sid = match s["session_id"].as_str().or_else(|| s["id"].as_str()) {
-            Some(s) => s,
-            None => return false,
-        };
-        active_sessions
-            .get(aid)
-            .is_some_and(|set| set.contains(sid))
-    };
-
-    if json {
-        // Annotate each session with `state` so JSON consumers see the same
-        // signal as the table renderer.
-        if let Some(arr) = session_arr_owned.as_ref() {
-            let annotated: Vec<serde_json::Value> = arr
-                .iter()
-                .filter(|s| !active_only || is_running(s))
-                .map(|s| {
-                    let mut out = s.clone();
-                    out["state"] = serde_json::Value::String(
-                        if is_running(s) { "running" } else { "idle" }.into(),
-                    );
-                    out
-                })
-                .collect();
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&annotated).unwrap_or_default()
-            );
-        } else {
-            println!(
-                "{}",
-                serde_json::to_string_pretty(&body).unwrap_or_default()
-            );
-        }
-        return;
-    }
-    if let Some(arr) = session_arr_owned.as_ref() {
-        let filtered: Vec<&serde_json::Value> = arr
-            .iter()
-            .filter(|s| !active_only || is_running(s))
-            .collect();
-        if filtered.is_empty() {
-            if active_only {
-                println!("No active sessions.");
-            } else {
-                println!("No sessions found.");
-            }
-            return;
-        }
-        let mut t = crate::table::Table::new(&["ID", "AGENT", "MSGS", "STATE", "LAST ACTIVE"]);
-        for s in filtered {
-            let state = if is_running(s) { "running" } else { "idle" };
-            let agent_id = s["agent_id"].as_str().unwrap_or("");
-            let agent_col = if agent_id.len() > 16 {
-                &agent_id[..16]
-            } else if agent_id.is_empty() {
-                s["agent_name"].as_str().unwrap_or("?")
-            } else {
-                agent_id
-            };
-            t.add_row(&[
-                s["session_id"]
-                    .as_str()
-                    .or_else(|| s["id"].as_str())
-                    .unwrap_or("?"),
-                agent_col,
-                &s["message_count"].as_u64().unwrap_or(0).to_string(),
-                state,
-                s["created_at"]
-                    .as_str()
-                    .or_else(|| s["last_active"].as_str())
-                    .unwrap_or("?"),
-            ]);
-        }
-        t.print();
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    }
-}
-
 fn cmd_security_status(json: bool) {
     let base = require_daemon("security status");
     let client = daemon_client();
@@ -7682,33 +4652,6 @@ fn cmd_webhooks_test(id: &str) {
             "webhook-test-failed",
             &[("error", body["error"].as_str().unwrap_or("?"))],
         ));
-    }
-}
-
-fn cmd_message(agent: &str, text: &str, json: bool, incognito: bool) {
-    let base = require_daemon("message");
-    let agent_id = resolve_agent_id(&base, agent);
-    let client = daemon_client();
-    let body = daemon_json(
-        client
-            .post(format!("{base}/api/agents/{agent_id}/message"))
-            .json(&serde_json::json!({"message": text, "incognito": incognito}))
-            .send(),
-    );
-    if json {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
-    } else if let Some(reply) = body["reply"].as_str() {
-        println!("{reply}");
-    } else if let Some(reply) = body["response"].as_str() {
-        println!("{reply}");
-    } else {
-        println!(
-            "{}",
-            serde_json::to_string_pretty(&body).unwrap_or_default()
-        );
     }
 }
